refactor: organize project
