+++ src/binding.cpp
@@ -1,10 +1,10 @@
-#include "binding.h"
+#include <binding.hpp>
+#include <utils.hpp>
 
 #include <iostream>
 #include <optional>
 #include <vector>
 #include <cstring>
-#include <iomanip>
 
 void* LoadModel(const char *modelPath, int numberGpuLayers)
 {
@@ -17,17 +17,6 @@ void* LoadModel(const char *modelPath, int numberGpuLayers)
     llama_add_eos_token(model);
 
     return model;
-}
-
-void PrintPerformanceInfo(const llama_context* context) {
-    const auto data = llama_perf_context(context);
-
-    float prompt_tok_per_sec = 1e3 / data.t_p_eval_ms * data.n_p_eval;
-    float gen_tok_per_sec = 1e3 / data.t_eval_ms * data.n_eval;
-
-    std::cout << "\n\n" << std::fixed << std::setprecision(2)
-              << "Prompt Processing: " << prompt_tok_per_sec << " tok/s, "
-              << "Text Generation: " << gen_tok_per_sec << " tok/s" << "\n" << std::endl;
 }
 
 struct ReadbackBuffer
